Added virtual list strange ratios
