+++ tests/dummy/app/routes/mobile-list.js
@@ -2,7 +2,5 @@ import Ember from 'ember';
 import makeModel from '../utils/make-model';
 
 export default Ember.Route.extend({
-  model: function() {
-    return makeModel();
-  }
+  model: makeModel()
 });
