+++ tests/dummy/app/routes/virtual-strange-ratios.js
@@ -2,5 +2,5 @@ import Ember from 'ember';
 import makeModel from '../utils/make-model';
 
 export default Ember.Route.extend({
-  model: makeModel(1000, 'smallImages')
+  model: makeModel(15, 'strangeRatios')
 });
