+++ tests/dummy/app/templates/application.hbs
@@ -9,5 +9,6 @@
     <li>{{#link-to 'multi-height-virtual-list'}}Multi Height - Virtual{{/link-to}}</li>
     <li>{{#link-to 'pull-to-refresh-list'}}Pull to refresh{{/link-to}}</li>
     <li>{{#link-to 'virtual'}}Virtual{{/link-to}}</li>
+    <li>{{#link-to 'virtual-strange-ratios'}}Virtual Strange Ratios{{/link-to}}</li>
 </ul>
 {{outlet}}
