+++ tests/dummy/app/utils/fixtures.js
@@ -1,0 +1,34 @@
+export var types = [
+  {id:  1, type: "cat",   name: "Andrew"},
+  {id:  2, type: "cat",   name: "Andrew"},
+  {id:  3, type: "cat",   name: "Bruce"},
+  {id:  4, type: "other", name: "Xbar"},
+  {id:  5, type: "dog",   name: "Caroline"},
+  {id:  6, type: "cat",   name: "David"},
+  {id:  7, type: "other", name: "Xbar"},
+  {id:  8, type: "other", name: "Xbar"},
+  {id:  9, type: "dog",   name: "Edward"},
+  {id: 10, type: "dog",   name: "Francis"},
+  {id: 11, type: "dog",   name: "George"},
+  {id: 12, type: "other", name: "Xbar"},
+  {id: 13, type: "dog",   name: "Harry"},
+  {id: 14, type: "cat",   name: "Ingrid"},
+  {id: 15, type: "other", name: "Xbar"},
+  {id: 16, type: "cat",   name: "Jenn"},
+  {id: 17, type: "cat",   name: "Kelly"},
+  {id: 18, type: "other", name: "Xbar"},
+  {id: 19, type: "other", name: "Xbar"},
+  {id: 20, type: "cat",   name: "Larry"},
+  {id: 21, type: "other", name: "Xbar"},
+  {id: 22, type: "cat",   name: "Manny"},
+  {id: 23, type: "dog",   name: "Nathan"},
+  {id: 24, type: "cat",   name: "Ophelia"},
+  {id: 25, type: "dog",   name: "Patrick"},
+  {id: 26, type: "other", name: "Xbar"},
+  {id: 27, type: "other", name: "Xbar"},
+  {id: 28, type: "other", name: "Xbar"},
+  {id: 29, type: "other", name: "Xbar"},
+  {id: 30, type: "other", name: "Xbar"},
+  {id: 31, type: "cat",   name: "Quincy"},
+  {id: 32, type: "dog",   name: "Roger"},
+  ];
