+++ tests/dummy/app/utils/images.js
@@ -1,0 +1,43 @@
+var images = [
+  'images/ebryn.jpg',
+  'images/iterzic.jpg',
+  'images/kselden.jpg',
+  'images/machty.jpg',
+  'images/rwjblue.jpg',
+  'images/stefanpenner.jpg',
+  'images/tomdale.jpg',
+  'images/trek.jpg',
+  'images/wagenet.jpg',
+  'images/wycats.jpg'
+];
+
+var smallImages = [
+  'images/small/Ba_Gua_Feng-Shui-Mirror.gif',
+  'images/small/Bonsai.gif',
+  'images/small/Chouchin_Reinensai_Lantern.gif',
+  'images/small/Chouchin_Kuroshiro_Lantern_.gif',
+  'images/small/Chouchin_Shinku_Lantern.gif',
+  'images/small/Fuurin_Glass_Wind_Chime.gif',
+  'images/small/Geta_Wooden_Sandal_.gif',
+  'images/small/Gunsen_Fan_.gif',
+  'images/small/iChing_Kouka_Heads-Coin.gif',
+  'images/small/iChing_Kouka_Tails_Coin.gif',
+  'images/small/Ishidourou_Snow_Lantern.gif',
+  'images/small/Kakejiku_Hanging_Scroll.gif',
+  'images/small/Katana_and_Sheath.gif',
+  'images/small/Kimono_Buru_Blue.gif',
+  'images/small/Kimono_Chairo_Tan.gif',
+  'images/small/Koi.gif',
+  'images/small/Shamisen.gif',
+  'images/small/Shodou_Calligraphy.gif',
+  'images/small/Torii.gif',
+  'images/small/Tsukubai_Water_Basin.gif'
+];
+
+var strangeRatios = [
+  'images/strange-ratios/horizontal-rectangle.jpg',
+  'images/strange-ratios/square.jpg',
+  'images/strange-ratios/vertical-rectangle.jpg'
+];
+
+export default { images, smallImages, strangeRatios };
